requirements.txt
Django==3.2.8
djongo==1.3.6
django-allauth
dnspython
djangorestframework
drf-yasg
